build for arm cleanup

+++ scripts/linux/build_iconv.sh
@@ -6,17 +6,22 @@ export ICONV_FILE_PATH=$WORKDIR/$ICONV_FILENAME
 export ICONV_SRC_DIR=$WORKDIR/libiconv-1.16
 ICONV_SHA256="e6a1b1b589654277ee790cce3734f07876ac4ccfaecbee8afa0b649cf529cc04"
 
-curl http://ftp.gnu.org/pub/gnu/libiconv/$ICONV_FILENAME -o $ICONV_FILE_PATH
+if [ ! -e "$ICONV_FILE_PATH" ]; then
+  curl http://ftp.gnu.org/pub/gnu/libiconv/$ICONV_FILENAME -o $ICONV_FILE_PATH
+fi
+
 echo $ICONV_SHA256 $ICONV_FILE_PATH | sha256sum -c - || exit 1
 
-for arch in x86_64
+for arch in $TYPES_OF_BUILD
 do
 
 PREFIX=${WORKDIR}/prefix_${arch}
 
 case $arch in
-	"aarch"	)
+	"x86_64"	)
         HOST="x86_64-linux-gnu";;
+	"aarch64"	)
+        HOST="aarch64-linux-gnu";;
 	*		)
 		HOST="x86_64-linux-gnu";;
 esac 
@@ -25,7 +30,7 @@ cd $WORKDIR
 rm -rf $ICONV_SRC_DIR
 tar -xzf $ICONV_FILE_PATH -C $WORKDIR
 cd $ICONV_SRC_DIR
-./configure --build=x86_64-linux-gnu --host=${HOST} --prefix=${PREFIX} --disable-rpath
+./configure --build=x${HOST} --host=${HOST} --prefix=${PREFIX} --disable-rpath
 make -j$THREADS
 make install
